Redux holds auth state, login/logout hint
Modified login/logout hint that displays with Navigation to depends is user logged in or not.

Added auths to store with auths reducers, which holds information about auth state, and data of actually logged if any exists. Additionally firebase is indicating SIGN_IN and SIGN_OUT actions in auths reducer by onAuthStateChanged() function.